% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catp.R
\name{catp}
\alias{catp}
\title{Print a vector of character strings to a connection.}
\usage{
catp(str, conn, sep = "\\n")
}
\arguments{
\item{str}{- vector of character strings}

\item{conn}{- connection to print to}

\item{sep}{- separator to use between vector elements}
}
\description{
Function to print a vector of character strings to a connection.
}
\details{
Shortcut for \code{cat(str,sep=sep,file=conn)}.
}
