% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberFunctions.R
\name{crabYear}
\alias{crabYear}
\title{Convert years to crab year strings}
\usage{
crabYear(year)
}
\arguments{
\item{year}{- year (numeric or character) to convert}
}
\value{
character vector with years in format YYYY/YY+1.
}
\description{
Function to convert year to crab year string.
}
\details{
crab year is in format YYYY/YY+1.
}
