% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escapeChars.R
\name{escapeChars}
\alias{escapeChars}
\title{Escape special characters in character vector.}
\usage{
escapeChars(x, env = "latex")
}
\arguments{
\item{x}{- character vector in which to escape special characters}

\item{env}{- environment for escaped characters}
}
\value{
character vector with special characters escaped.
}
\description{
Function to escape special characters in character vector.
}
\details{
Currently escapes the following characters: &, %.
}
