% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figInfo.ParseFromMarkdownFile.R
\name{figInfo.ParseFromMarkdownFile}
\alias{figInfo.ParseFromMarkdownFile}
\title{Parse a markdown (.md) file to extract information on figures by document section}
\usage{
figInfo.ParseFromMarkdownFile(fn, verbose = FALSE)
}
\arguments{
\item{fn}{- the markdown (.md) filename, including path}
}
\value{
a figInfo tibble (see [figInfo()])
}
\description{
Function to parse a markdown (.md) file to extract information on figures by document section.
}
\details{
The markdown file is assumed to be in markdown format for Latex.
}
