% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figInfo.R
\name{figInfo}
\alias{figInfo}
\title{Create a figInfo object (a figure information tibble), or append figure information to one}
\usage{
figInfo(
  .x,
  label = "",
  path = "",
  fn = "",
  width = NA_real_,
  height = NA_real_,
  orientation = c("portrait", "landscape"),
  caption = ""
)
}
\arguments{
\item{.x}{\itemize{
\item NULL, or an existing figure information dataframe (see \link{@details})
}}

\item{label}{\itemize{
\item the latex label for the figure
}}

\item{path}{\itemize{
\item path to file (not including filename)
}}

\item{fn}{\itemize{
\item the filename for a figure or image
}}

\item{width}{\itemize{
\item figure width (in inches)
}}

\item{height}{\itemize{
\item figure height (in inches)
}}

\item{orientation}{\itemize{
\item the orientation for the figure ("portrait" or "landscape")
}}

\item{caption}{\itemize{
\item the caption for the figure
}}
}
\value{
A figure information tibble (with a single row or the input dataframe with an added row of information).
}
\description{
Function to create a figInfo object (a figure information tibble), or append figure information to one.
}
\details{
A figure information dataframe has columns label, path, fn, width, height, orientation, and caption.
}
