% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/figInfo.ReadCSV.R
\name{figInfo.ReadCSV}
\alias{figInfo.ReadCSV}
\title{Read figInfo from a csv file and add it to an existing figInfo object}
\usage{
figInfo.ReadCSV(allfi, fncsv)
}
\arguments{
\item{allfi}{\itemize{
\item existing figInfo object, or NULL
}}

\item{fncsv}{\itemize{
\item name of csv file in figInfo format, or a dataframe in figInfo format
}}
}
\value{
a dataframe in figInfo format with information from the csv file appended
}
\description{
Function to read figInfo from a csv file and add it to an existing figInfo object.
}
\details{
Filenames in the csv file ending in a blank are assumed to be pdf files (under the assumption
the csv file was crated by \code{\link[=parseFigInfoFromMarkdownFile]{parseFigInfoFromMarkdownFile()}}. See \code{\link[=figInfo]{figInfo()}} for format information on returned dataframe.
=======
% Please edit documentation in R/figInfo.R
\name{figInfo.ReadCSV}
\alias{figInfo.ReadCSV}
\title{Read figInfo from a csv file and add it to an eexisting figInfo object}
\usage{
figInfo.ReadCSV(allfi, fncsv)
}
\description{
Read figInfo from a csv file and add it to an eexisting figInfo object
>>>>>>> 2114d8f49972ce60383a86af1fd43943cd47b76c
}
