% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureInfo.R
\name{figureInfo}
\alias{figureInfo}
\title{Create a figure information dataframe, or append figure information to one}
\usage{
figureInfo(
  .x,
  caption = "",
  label = "",
  file = "",
  width = 6.5,
  height = 6.5,
  orientation = c("portrait", "landscape")
)
}
\arguments{
\item{.x}{- NULL, or an existing figure information dataframe (see [@details])}

\item{label}{- the latex label for the figure}

\item{file}{- the path to a file containing a figure or image}

\item{width}{- figure width (in inches)}

\item{height}{- figure height (in inches)}

\item{orientation}{- the orientation for the figure ("portrait" or "landscape")}

\item{cap}{- the caption for the figure}
}
\value{
A figure information dataframe (with a single row or the input dataframe with an added row of information).
}
\description{
Function to create a figure information dataframe, or append figure information to one.
}
\details{
A figure information dataframe has columns label, file, caption, width, height, and orientation.
}
