% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixColLabelsForRowFactors.R
\name{fixColLabelsForRowFactors}
\alias{fixColLabelsForRowFactors}
\title{Fix column labels for row factors in a \code{tabular} table}
\usage{
fixColLabelsForRowFactors(tbr, labelsForRowFactors, dim = 2, verbose = FALSE)
}
\arguments{
\item{tbr}{\itemize{
\item tabular table
}}

\item{labelsForRowFactors}{\itemize{
\item character vector of desired column names for row factors
}}

\item{dim}{\itemize{
\item dimension of the tbr's dimnames attribute in which the row labels reside
}}

\item{-}{verbose - flag (T/F) to print diagnostic information}
}
\value{
\code{tbr}, with labels for row factors updated to the input \code{labelsForRowFactors}.
}
\description{
Function to fix column labels for row factors in a \code{tabular} table
}
\details{
The \code{\link[tables:tabular]{tables::tabular()}} function seems to have a problem with row factors that have one value, and
that is that the resulting tabular object is lacking column headings for some (most?) of those factors--something
which the conversion to pdf does not like. This function gives a quick way for assiging the "correct" headings (or new ones,
for that matter) to a tabular object. At this point, the tables I've dealt with have all had the row factor column headings in the 2nd dimension
of the rorwLabel object's "dimnames" attribute, but I suspect this may vary with the dimension of the factor crossing (hence \code{dim},
which default is 2).

Note also that this function may provide an easy way to fix collisions with latex constraints on column headings (it doesn't like underscores in headings, for one),
at least for those reflecting row factors.
}
