% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAspectRatio.R
\name{getAspectRatio}
\alias{getAspectRatio}
\title{Get the aspect ratio of an image}
\usage{
getAspectRatio(img)
}
\arguments{
\item{img}{\itemize{
\item an image object or a filename (as a string)
}}
}
\value{
the aspect ratio (height/width)
}
\description{
Function to get the aspect ratio of an image.
}
\details{
Uses \code{\link[=getImageDims]{getImageDims()}} to obtain the aspect ratio.
}
