% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImageDims.R
\name{getImageDims}
\alias{getImageDims}
\title{Get the dimensions of an image}
\usage{
getImageDims(img)
}
\arguments{
\item{img}{\itemize{
\item an image object or a filename (as a string)
}}
}
\value{
list with elements w (width in inches), h (height in inches), and asp (aspect=h/w)
}
\description{
Function to get the dimensions of an image.
}
\details{
Uses \code{\link[magick:editing]{magick::image_read()}} to read an image file (if \code{img} is a character string) and
\code{\link[magick:editing]{magick::image_read()}} to obtain size information on the image object. Width/height units are given in
inches.
}
