% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertImage.R
\name{insertImage}
\alias{insertImage}
\title{Create the latex to insert an image into a latex document}
\usage{
insertImage(fn, cap, lbl = cap, width = 6.5, height = 6, align = "ht!")
}
\arguments{
\item{fn}{- filename (including path) of image to insert}

\item{cap}{- the caption for the image}

\item{lbl}{- the latex label for the image}

\item{width}{- figure width (in inches)}

\item{height}{- figure height (in inches)}

\item{align}{- character string indicating latex alignment}
}
\value{
nothing.
}
\description{
Function create the latex to insert an image into a latex document.
}
\details{
Intended for use in a pdf-bound RMD script to create the latex
code to insert an image into a pdf document.
}
