% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberFunctions.R
\name{last}
\alias{last}
\title{Get the last n values of a vector}
\usage{
last(x, n)
}
\arguments{
\item{x}{- vector}
}
\value{
a vector of length \code{n}
}
\description{
Function to get the last \code{n} values of a vector.
}
\details{
none.
}
