% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberFunctions.R
\name{numKG2T}
\alias{numKG2T}
\title{Apply a pretty format to numbers in kg, converted to t}
\usage{
numKG2T(x)
}
\arguments{
\item{x}{- numeric or character vector in kg}
}
\value{
- character vector in t, formatted using [numT()]
}
\description{
Function to apply a pretty format to numbers in kg, converted to t.
}
\details{
- Uses [stripNumSyms()] (if necessary) and [numT()].
}
