% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberFunctions.R
\name{numT}
\alias{numT}
\title{Apply a pretty format for numbers in metric tons}
\usage{
numT(x)
}
\arguments{
\item{x}{- vector (numeric or character) of numbers to apply pretty format to}
}
\value{
character vector
}
\description{
Function to apply a pretty format for numbers in metric tons.
}
\details{
Values >= 1000 are rounded to single ones, values > 100 are formatted
with 2 decimal places, values >0 are formatted with 3 decimal places.
}
