% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseFigInfoFromMarkdownFile.R
\name{parseFigInfoFromMarkdownFile}
\alias{parseFigInfoFromMarkdownFile}
\title{Parse a markdown (.md) file to extract information on figures bby document section}
\usage{
parseFigInfoFromMarkdownFile(fn, verbose = FALSE)
}
\arguments{
\item{fn}{- the markdown (.md) filename, including path}
}
\value{
list with elements level, section, and figInfo (a nested list, or an empty list)
}
\description{
Funtion to parse a markdown (.md) file to extract information on figures by document section.
}
\details{
The markdown file is assumed to be in markdown format for Latex.
}
