% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseFigInfoText.R
\name{parseFigInfoText}
\alias{parseFigInfoText}
\title{Parse figure info text from a latex md file}
\usage{
parseFigInfoText(figtxt, verbose = FALSE)
}
\arguments{
\item{figtxt}{the figure text (i.e., text between \\begin\{figure\} and \\end\{figure\} delimiters)}

\item{verbose}{- flag (T/F) to print diagnostic information}
}
\value{
a figInfo object (a tibble with columns label, path, fn, width, height, orientation and caption)
}
\description{
Function to parse figure info text from a latex md file.
}
\details{
Parsing assumes the input file is in markdown format for Latex.
}
