% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processTableInfo.R
\name{processTableInfo}
\alias{processTableInfo}
\title{Process a table info object or file}
\usage{
processTableInfo(tblInfo, verbose = FALSE)
}
\arguments{
\item{tblInfo}{\itemize{
\item tblInfo dataframe or filename
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
nothing
}
\description{
Function to process a table info object or file.
}
\details{
nothing.
}
