% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeLabels.R
\name{sanitizeLabels}
\alias{sanitizeLabels}
\title{Sanitize labels by replacing illegal characters with "."}
\usage{
sanitizeLabels(x, env = "latex")
}
\arguments{
\item{x}{- character vector to sanitize}

\item{env}{- environment for sanitizing characters (default="latex")}
}
\value{
character vector with sanitized labels.
}
\description{
Function to sanitize labels by replacing illegal characters with ".".
}
\details{
Currently sanitizes the following characters: "&", "%", "_" in
a latex environment.
}
