% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberFunctions.R
\name{t2MLB}
\alias{t2MLB}
\title{Convert character vector from metric tons to numeric vector in millions of lbs}
\usage{
t2MLB(x)
}
\arguments{
\item{x}{- numeric of character vector of numbers in metric tons}
}
\value{
numeric vector in millions of lbs
}
\description{
Function to convert character string from metric tons to millions of lbs.
}
