% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableInfo.R
\name{tableInfo}
\alias{tableInfo}
\title{Create a table information dataframe, or append table information to one}
\usage{
tableInfo(
  .x,
  caption = "",
  label = "",
  path = "",
  fn = "",
  typesize = c("normal", "small"),
  orientation = c("portrait", "landscape"),
  type = c("latex", "pdf"),
  latex = ""
)
}
\arguments{
\item{.x}{- NULL, or an existing table information dataframe (see [Details])}

\item{label}{- the latex label for the table}

\item{path}{- the path to a file containing a latex table}

\item{fn}{- the name of the file containing a latex table}

\item{typesize}{- the type size for the table ("normal" or "small")}

\item{orientation}{- the orientation for the table ("portrait" or "landscape")}

\item{type}{- table format ("latex" or "pdf")}

\item{latex}{- latex-formatted table as character string or ""}

\item{cap}{- the caption for the table}
}
\value{
A tableInfo tibble (with a single row or the input dataframe with an added row of information).
}
\description{
Function to create a table information dataframe, or append table information to one.
}
\details{
A tableInfo tibble has columns label, path, fn, caption, typesize, orientation, (format) type, and latex.
}
