% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableInfo.WriteToMarkdown.R
\name{tableInfo.WriteToMarkdown}
\alias{tableInfo.WriteToMarkdown}
\title{Process a tableInfo object or file for inclusion in a markdown file}
\usage{
tableInfo.WriteToMarkdown(
  tblInfo,
  env = "latex",
  includeLabels = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{tblInfo}{\itemize{
\item a tableInfo object (dataframe) or filename for csv file in tableInfo format
}}

\item{env}{\itemize{
\item "environment" in which to write markdown (default="latex")
}}

\item{includeLabels}{\itemize{
\item include label caption (default=FALSE)
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
nothing
}
\description{
Function to process a tableInfo object or file for inclusion in a markdown file.
}
\details{
Creates text in markdown format to insert tables specified in a tableInfo object
into a markdown file. Handles tables represented as pdf files or as latex code in text files.
See \code{\link[=tableInfo]{tableInfo()}}.

The \code{env} parameter is used to specify the output format for escaping characters in captions.
See \code{\link[=escapeChars]{escapeChars()}}.

Setting \code{includeLabels} adds the label to the caption to make referencing easier (only works
for inserted pdf tables, not inserted text tables).
}
