% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableInfo.WriteToRMD.R
\name{tableInfo.WriteToRMD}
\alias{tableInfo.WriteToRMD}
\title{Process a tableInfo object or file for inclusion in a RMD file}
\usage{
tableInfo.WriteToRMD(tblInfo, docformat = "latex", verbose = FALSE)
}
\arguments{
\item{tblInfo}{\itemize{
\item a tableInfo object (dataframe) or filename for csv file in tableInfo format
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
nothing
}
\description{
Function to process a tableInfo object or file for inclusion in a markdown file.
}
\details{
Creates text in markdown format to insert tables in a tableInfo object
into a markdown file. Handles tables represented as pdf files or as latex code in text files.
See \code{\link[=tableInfo]{tableInfo()}}.
}
