% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTabularToLatex.R
\name{writeTabularToLatex}
\alias{writeTabularToLatex}
\title{Write a tabular object in latex format to a connection}
\usage{
writeTabularToLatex(
  tbl,
  cap,
  conn = "",
  label = NULL,
  justification = "l",
  type = c("", "longtable"),
  size = c("", "small")
)
}
\arguments{
\item{tbl}{- a tabular object}

\item{cap}{- caption}

\item{conn}{- connection to write to}

\item{label}{- label for latex}

\item{justification}{- character indicating table justification ("l","r")}

\item{type}{- table type ("","longtable")}

\item{size}{- font size ("","small")}
}
\value{
nothing.
}
\description{
Function to write a tabular object in latex format to a connection.
}
\details{
If \code{conn} is not "", the connection is opened for writing to text.
Uses [tables::toLatex()] to create the latex representation of the tabular object.
}
