% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figInfo.R
\name{figInfo.ReadCSV}
\alias{figInfo.ReadCSV}
\title{Read figInfo from a csv file and add it to an eexisting figInfo object}
\usage{
figInfo.ReadCSV(allfi, fncsv)
}
\description{
Read figInfo from a csv file and add it to an eexisting figInfo object
}
