% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figInfo.ReadCSV.R
\name{figInfo.ReadCSV}
\alias{figInfo.ReadCSV}
\title{Read figInfo from a csv file and add it to an existing figInfo object}
\usage{
figInfo.ReadCSV(allfi, fncsv)
}
\arguments{
\item{allfi}{\itemize{
\item existing figInfo object, or NULL
}}

\item{fncsv}{\itemize{
\item name of csv file in figInfo format, or a dataframe in figInfo format
}}
}
\value{
a dataframe in figInfo format with information from the csv file appended
}
\description{
Function to read figInfo from a csv file and add it to an existing figInfo object.
}
\details{
Filenames in the csv file ending in a blank are assumed to be pdf files (under the assumption
the csv file was crated by \code{\link[=figInfo.ParseFromMarkdownFile]{figInfo.ParseFromMarkdownFile()}}. See \code{\link[=figInfo]{figInfo()}} for format information on returned dataframe.
}
