% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberFunctions.R
\name{numKT}
\alias{numKT}
\title{Apply a pretty format for numbers in 1,000's of metric tons}
\usage{
numKT(x)
}
\arguments{
\item{x}{- vector (numeric or character) of numbers to apply pretty format to}
}
\value{
character vector
}
\description{
Function to apply a pretty format for numbers in 1,000's of metric tons.
}
\details{
Values >= 1000 thousand t are rounded to single ones, values > 100 thousand t are formatted
with 1 decimal place, values > 10 are formatted with 2 decimal places, values > 0
are formatted with 3 decimal places.
}
