% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseFigInfoText.R
\name{parseFigInfoText}
\alias{parseFigInfoText}
\title{Parse figure info text from a latex md file}
\usage{
parseFigInfoText(figtxt)
}
\arguments{
\item{figtxt}{the figure text (i.e., text between \\begin\{figure\} and \\end\{figure\} delimiters)}
}
\value{
list with elements label, path, width, height, and caption
}
\description{
Function to parse figure info text from a latex md file.
}
\details{
Parsing assumes the input file is in markdown format for Latex.
}
