% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTabularToLatex.R
\name{writeTabularToLatex}
\alias{writeTabularToLatex}
\title{Write a tabular object in latex format to a connection}
\usage{
writeTabularToLatex(
  tbl,
  cap,
  conn = "",
  label = NULL,
  justification = "l",
  type = c("", "longtable"),
  size = c("", "small")
)
}
\arguments{
\item{tbl}{\itemize{
\item a tabular object
}}

\item{cap}{\itemize{
\item table caption
}}

\item{conn}{\itemize{
\item connection to write to
}}

\item{label}{\itemize{
\item label for referencing
}}

\item{justification}{\itemize{
\item character indicating table justification ("l","r")
}}

\item{type}{\itemize{
\item table type ("","longtable")
}}

\item{size}{\itemize{
\item font size ("","small")
}}
}
\value{
string version of table in latex format (invisibly)
}
\description{
Function to write a tabular object in latex format to a connection.
}
\details{
If \code{conn} is "", the created latex string is written to the standard output connection (in most cases, the console).
If \code{conn} is not "", the connection is opened for writing to text and the string is written to it. If \code{conn=NULL}, nothing is written.
In all cases, the string is returned invisibly
to the calling function.

This uses \code{\link[tables:latex.tabular]{tables::toLatex()}} to create the latex representation of the tabular object.
}
